Changes:
User input checks improved, a user's input should not break the program under any circumstance now.

Javascript was ruled out for making this program a web app (it lacks true multi-dimensional arrays).

The web app version is now being made with Wt for c++,  allowing the reuse of much of this code.